refactor(admin): move comp into own files
