+++ adminSiteClient/ChartList.tsx
@@ -1,20 +1,13 @@
 import React from "react"
 import { observer } from "mobx-react"
-import { action, runInAction, observable } from "mobx"
-import * as lodash from "lodash"
-
-import { Link } from "./Link.js"
+import { runInAction, observable } from "mobx"
 import { Tag } from "./TagBadge.js"
 import { bind } from "decko"
-import { EditableTags, Timeago } from "./Forms.js"
 import { AdminAppContext, AdminAppContextType } from "./AdminAppContext.js"
-import {
-    BAKED_GRAPHER_EXPORTS_BASE_URL,
-    BAKED_GRAPHER_URL,
-} from "../settings/clientSettings.js"
 import { ChartTypeName, GrapherInterface } from "@ourworldindata/grapher"
 import { startCase } from "@ourworldindata/utils"
 import { References, getFullReferencesCount } from "./ChartEditor.js"
+import { ChartRow } from "./ChartRow.js"
 
 // These properties are coming from OldChart.ts
 export interface ChartListItem {
@@ -37,129 +30,6 @@ export interface ChartListItem {
     isExplorable: boolean
 
     tags: Tag[]
-}
-
-function showChartType(chart: ChartListItem) {
-    const chartType = chart.type ?? ChartTypeName.LineChart
-    const displayType = ChartTypeName[chartType]
-        ? startCase(ChartTypeName[chartType])
-        : "Unknown"
-
-    if (chart.tab === "map") {
-        if (chart.hasChartTab) return `Map + ${displayType}`
-        else return "Map"
-    } else {
-        if (chart.hasMapTab) return `${displayType} + Map`
-        else return displayType
-    }
-}
-
-@observer
-class ChartRow extends React.Component<{
-    chart: ChartListItem
-    searchHighlight?: (text: string) => string | JSX.Element
-    availableTags: Tag[]
-    onDelete: (chart: ChartListItem) => void
-}> {
-    static contextType = AdminAppContext
-    context!: AdminAppContextType
-
-    async saveTags(tags: Tag[]) {
-        const { chart } = this.props
-        const json = await this.context.admin.requestJSON(
-            `/api/charts/${chart.id}/setTags`,
-            { tags },
-            "POST"
-        )
-        if (json.success) {
-            runInAction(() => (chart.tags = tags))
-        }
-    }
-
-    @action.bound onSaveTags(tags: Tag[]) {
-        this.saveTags(tags)
-    }
-
-    render() {
-        const { chart, searchHighlight, availableTags } = this.props
-
-        const highlight = searchHighlight || lodash.identity
-
-        return (
-            <tr>
-                <td style={{ minWidth: "140px", width: "12.5%" }}>
-                    {chart.isPublished && (
-                        <a href={`${BAKED_GRAPHER_URL}/${chart.slug}`}>
-                            <img
-                                src={`${BAKED_GRAPHER_EXPORTS_BASE_URL}/${chart.slug}.svg`}
-                                className="chartPreview"
-                            />
-                        </a>
-                    )}
-                </td>
-                <td style={{ minWidth: "180px" }}>
-                    {chart.isPublished ? (
-                        <a href={`${BAKED_GRAPHER_URL}/${chart.slug}`}>
-                            {highlight(chart.title ?? "")}
-                        </a>
-                    ) : (
-                        <span>
-                            <span style={{ color: "red" }}>Draft: </span>{" "}
-                            {highlight(chart.title ?? "")}
-                        </span>
-                    )}{" "}
-                    {chart.variantName ? (
-                        <span style={{ color: "#aaa" }}>
-                            ({highlight(chart.variantName)})
-                        </span>
-                    ) : undefined}
-                    {chart.internalNotes && (
-                        <div className="internalNotes">
-                            {highlight(chart.internalNotes)}
-                        </div>
-                    )}
-                </td>
-                <td style={{ minWidth: "100px" }}>{chart.id}</td>
-                <td style={{ minWidth: "100px" }}>{showChartType(chart)}</td>
-                <td style={{ minWidth: "340px" }}>
-                    <EditableTags
-                        tags={chart.tags}
-                        suggestions={availableTags}
-                        onSave={this.onSaveTags}
-                        hasKeyChartSupport={true}
-                    />
-                </td>
-                <td>
-                    <Timeago
-                        time={chart.publishedAt}
-                        by={highlight(chart.publishedBy)}
-                    />
-                </td>
-                <td>
-                    <Timeago
-                        time={chart.lastEditedAt}
-                        by={highlight(chart.lastEditedBy)}
-                    />
-                </td>
-                <td>
-                    <Link
-                        to={`/charts/${chart.id}/edit`}
-                        className="btn btn-primary"
-                    >
-                        Edit
-                    </Link>
-                </td>
-                <td>
-                    <button
-                        className="btn btn-danger"
-                        onClick={() => this.props.onDelete(chart)}
-                    >
-                        Delete
-                    </button>
-                </td>
-            </tr>
-        )
-    }
 }
 
 @observer
@@ -261,5 +131,20 @@ export class ChartList extends React.Component<{
                 </tbody>
             </table>
         )
+    }
+}
+
+export function showChartType(chart: ChartListItem) {
+    const chartType = chart.type ?? ChartTypeName.LineChart
+    const displayType = ChartTypeName[chartType]
+        ? startCase(ChartTypeName[chartType])
+        : "Unknown"
+
+    if (chart.tab === "map") {
+        if (chart.hasChartTab) return `Map + ${displayType}`
+        else return "Map"
+    } else {
+        if (chart.hasMapTab) return `${displayType} + Map`
+        else return displayType
     }
 }
