+++ adminSiteClient/PostsIndexPage.tsx
@@ -6,7 +6,8 @@ import * as lodash from "lodash"
 
 import { highlight as fuzzyHighlight } from "@ourworldindata/grapher"
 import { AdminLayout } from "./AdminLayout.js"
-import { SearchField, FieldsRow, EditableTags, Timeago } from "./Forms.js"
+import { SearchField, FieldsRow, Timeago } from "./Forms.js"
+import { EditableTags } from "./EditableTags.js"
 import { AdminAppContext, AdminAppContextType } from "./AdminAppContext.js"
 import { WORDPRESS_URL } from "../settings/clientSettings.js"
 import { Tag } from "./TagBadge.js"
